apps/coprocessor/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.0
httpx==0.27.0
python-dotenv==1.0.1

# 开发工具
ruff==0.1.9

# 文件处理
aiofiles==24.1.0
Werkzeug==2.3.7

# 测试工具
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-mock==3.12.0
# pytest-cov==4.1.0  # 暂时注释掉，因为需要 SQLite3 支持

# ASR 服务
dashscope>=1.14.0

# OSS 服务
oss2>=2.18.0